+++ src/main/java/de/gematik/demis/pdfgen/FeatureFlags.java
@@ -37,4 +37,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 public class FeatureFlags {
   private boolean hospitalizationOrder;
   private boolean diseaseStrict;
+  private boolean pdfOptimization;
 }