+++ src/main/java/de/gematik/demis/pdfgen/receipt/common/model/subsection/Facility.java
@@ -41,4 +41,5 @@ public class Facility implements PrintableOrganization {
   private List<Telecom> telecoms;
   private NameDTO nameDTO;
   private String department;
+  private boolean reorderContactPersonText;
 }