+++ src/main/java/de/gematik/demis/pdfgen/receipt/laboratoryreport/model/LaboratoryReportTemplateDtoFactory.java
@@ -27,6 +27,7 @@ package de.gematik.demis.pdfgen.receipt.laboratoryreport.model;
  * #L%
  */
 
+import de.gematik.demis.pdfgen.FeatureFlags;
 import de.gematik.demis.pdfgen.receipt.common.model.section.AuthenticationFactory;
 import de.gematik.demis.pdfgen.receipt.common.model.section.MetadataFactory;
 import de.gematik.demis.pdfgen.receipt.common.model.section.NotificationFactory;
@@ -59,6 +60,7 @@ public class LaboratoryReportTemplateDtoFactory {
   private final AuthenticationFactory authenticationFactory;
   private final QrGenerator qrGenerator;
   private final WatermarkService watermarkService;
+  private final FeatureFlags featureFlags;
 
   public LaboratoryReportTemplateDto create(final Bundle bundle, final boolean qrCodeOnLastPage) {
     Optional<String> notificationIdOptional =
@@ -89,7 +91,8 @@ public class LaboratoryReportTemplateDtoFactory {
             .submitter(submitterFactory.create(bundle))
             .labReport(labReportFactory.create(bundle))
             .authentication(authenticationFactory.create(bundle))
-            .qrCodeOnLastPage(qrCodeOnLastPage);
+            .qrCodeOnLastPage(qrCodeOnLastPage)
+            .pdfOptimization(featureFlags.isPdfOptimization());
     watermarkService.getWatermarkBase64Image().ifPresent(builder::watermarkBase64Image);
     return builder.build();
   }