+++ src/test/java/de/gematik/demis/pdfgen/receipt/laboratoryreport/LaboratoryReportControllerIntegrationTest.java
@@ -29,10 +29,12 @@ package de.gematik.demis.pdfgen.receipt.laboratoryreport;
 
 import static de.gematik.demis.pdfgen.test.helper.FhirFactory.*;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
 import de.gematik.demis.fhir_ui_data_model_translation_service.wiremockfuts.WireMockFuts;
 import de.gematik.demis.fhir_ui_data_model_translation_service.wiremockfuts.laboratory.LaboratoryFeature;
+import de.gematik.demis.pdfgen.FeatureFlags;
 import de.gematik.demis.pdfgen.test.helper.PdfExtractorHelper;
 import java.awt.image.RenderedImage;
 import java.io.IOException;
@@ -42,24 +44,26 @@ import org.assertj.core.api.Assertions;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
+import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.TestPropertySources;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 import org.springframework.test.web.servlet.MockMvc;
+import org.wiremock.spring.EnableWireMock;
 
 @AutoConfigureMockMvc
-@AutoConfigureWireMock(port = 0)
+@EnableWireMock
 @SpringBootTest(
     properties = "server.port=10103",
     webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
@@ -76,18 +80,157 @@ class LaboratoryReportControllerIntegrationTest {
 
   @Autowired private MockMvc mockMvc;
 
+  @MockitoSpyBean FeatureFlags featureFlags;
+
   @BeforeEach
   void configureWireMockFuts() {
     new WireMockFuts().setDefaults().add(new LaboratoryFeature());
   }
 
-  @Test
-  void generatePdfFromDv2BundleJsonString_shouldRespond200WithPdf() throws Exception {
-    final MockHttpServletResponse response =
-        generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_JSON);
+  @Nested
+  class PdfOptimizationDisabled {
+    @BeforeEach
+    void setup() {
+      when(featureFlags.isPdfOptimization()).thenReturn(false);
+    }
 
-    validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
-    validateLabReportDv2Body(response, getReportDv2PdfText(""));
+    @Test
+    void generatePdfFromDv2BundleJsonString_shouldRespond200WithPdf() throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+      final String expectedNotification =
+          getExpectedNotificationPdfText("Meldungsverweis (Primärlabor) ABC123");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText("", expectedNotifierFacilityPdfText, null, expectedNotification));
+    }
+
+    @Test
+    void generatePdfFromDv2BundleJsonString_withoutRelatesTo_shouldNotHaveEntryInPdf()
+        throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_WITHOUT_RELATES_TO_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+
+      // FEATURE_FLAG_PDF_OPTIMIZATION is disabled -> no entry for initial notification id in pdf
+      final String expectedNotification = getExpectedNotificationPdfText("");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText("", expectedNotifierFacilityPdfText, null, expectedNotification));
+    }
+
+    @Test
+    void
+        generatePdfFromDv2BundleJsonString_withContactNameText_shouldHaveContactNameConcatenationAsContactPerson()
+            throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_WITH_CONTACT_TEXT_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      // FEATURE_FLAG_PDF_OPTIMIZATION is disabled -> entry "Kontaktperson" build from contact.name
+      // concatenation
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+      final String expectedSubmitterFacilityPdfText =
+          getExpectedSubmitterFacilityPdfText("Kontaktperson Dr Mila Careful");
+
+      final String expectedNotification =
+          getExpectedNotificationPdfText("Meldungsverweis (Primärlabor) ABC123");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText(
+              "",
+              expectedNotifierFacilityPdfText,
+              expectedSubmitterFacilityPdfText,
+              expectedNotification));
+    }
+  }
+
+  @Nested
+  class PdfOptimizationEnabled {
+    @BeforeEach
+    void setup() {
+      when(featureFlags.isPdfOptimization()).thenReturn(true);
+    }
+
+    @Test
+    void generatePdfFromDv2BundleJsonString_shouldRespond200WithPdf() throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+      final String expectedNotification =
+          getExpectedNotificationPdfText("Meldungsverweis (Initiale Meldungs-ID) ABC123");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText("", expectedNotifierFacilityPdfText, null, expectedNotification));
+    }
+
+    @Test
+    void generatePdfFromDv2BundleJsonString_withoutRelatesTo_shouldHaveEntryInPdf()
+        throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_WITHOUT_RELATES_TO_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+
+      // FEATURE_FLAG_PDF_OPTIMIZATION is enabled -> entry for initial notification id in pdf "Keine
+      // Angabe"
+      final String expectedNotification =
+          getExpectedNotificationPdfText("Meldungsverweis (Initiale Meldungs-ID) Keine Angabe");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText("", expectedNotifierFacilityPdfText, null, expectedNotification));
+    }
+
+    @Test
+    void
+        generatePdfFromDv2BundleJsonString_withContactNameText_shouldHaveContactNameTextAsContactPerson()
+            throws Exception {
+      final MockHttpServletResponse response =
+          generateLaboratoryPdf(LABORATORY_REPORT_BUNDLE_DV2_WITH_CONTACT_TEXT_JSON);
+
+      validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
+
+      // FEATURE_FLAG_PDF_OPTIMIZATION is enabled -> entry "Kontaktperson" build from
+      // contact.name.text
+      final String expectedNotifierFacilityPdfText =
+          getExpectedNotifierFacilityPdfText("Kontaktperson Dr. Adam Careful Notifier");
+      final String expectedSubmitterFacilityPdfText =
+          getExpectedSubmitterFacilityPdfText("Kontaktperson Dr. Mila Careful Submitter");
+
+      final String expectedNotification =
+          getExpectedNotificationPdfText("Meldungsverweis (Initiale Meldungs-ID) ABC123");
+
+      validateLabReportDv2Body(
+          response,
+          getReportDv2PdfText(
+              "",
+              expectedNotifierFacilityPdfText,
+              expectedSubmitterFacilityPdfText,
+              expectedNotification));
+    }
   }
 
   @Test
@@ -107,7 +250,15 @@ class LaboratoryReportControllerIntegrationTest {
         generateLaboratoryPdf(LABORATORY_NOTIFICATION_WITHOUT_SPECIMEN_STATUS);
 
     validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
-    String expected = getReportDv2PdfText("").replace("Probenmaterialstatus Verfügbar\n", "");
+
+    final String expectedNotifierFacilityPdfText =
+        getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+    final String expectedNotification =
+        getExpectedNotificationPdfText("Meldungsverweis (Initiale Meldungs-ID) ABC123");
+    String expected =
+        getReportDv2PdfText("", expectedNotifierFacilityPdfText, null, expectedNotification)
+            .replace("Probenmaterialstatus Verfügbar\n", "");
+
     validateLabReportDv2Body(response, expected);
   }
 
@@ -133,7 +284,16 @@ class LaboratoryReportControllerIntegrationTest {
     final MockHttpServletResponse response = generateLaboratoryPdf(pathToInputBundle);
 
     validateOkResponse(response, "Empfangsbestätigung Labormeldung - Maxime Mustermann.pdf");
-    validateLabReportDv2Body(response, getReportDv2PdfText(expectedProvenance));
+
+    final String expectedNotifierFacilityPdfText =
+        getExpectedNotifierFacilityPdfText("Kontaktperson Dr Adam Careful");
+    final String expectedNotification =
+        getExpectedNotificationPdfText("Meldungsverweis (Initiale Meldungs-ID) ABC123");
+
+    validateLabReportDv2Body(
+        response,
+        getReportDv2PdfText(
+            expectedProvenance, expectedNotifierFacilityPdfText, null, expectedNotification));
   }
 
   private @NotNull MockHttpServletResponse generateLaboratoryPdf(String bundle) throws Exception {
@@ -163,19 +323,20 @@ class LaboratoryReportControllerIntegrationTest {
     assertThat(response.getHeader("Content-Length")).isNotNull();
   }
 
-  String getReportDv2PdfText(String expectedProvenance) {
+  String getReportDv2PdfText(
+      String expectedProvenance,
+      final String expectedNotifierFacility,
+      final String expectedSubmitterFacility,
+      final String expectedNotification) {
     return """
     Empfangsbestätigung Labormeldung
     Vielen Dank für Ihre Meldung. Die Daten wurden an das zuständige Gesundheitsamt gemeldet.\s
     Ggf. wird man von dort Kontakt mit Ihnen aufnehmen, um weitere Daten zu ermitteln. Bitte\s
     speichern Sie die Meldungsquittung datenschutzrechtlich sicher ab, da diese personenbezogene\s
     Daten enthält.
-    Meldevorgangs-ID a5e00874-bb26-45ac-8eea-0bde76456703
-    Zeitpunkt des Eingangs 24.10.2023 09:06
-    Meldungs-ID e8d8cc43-32c2-4f93-8eaf-b2f3e6deb2a9
-    Meldungsstatus Final
-    Meldungserstellung/-änderung 04.03.2021 20:16
-    Meldungsverweis (Primärlabor) ABC123
+    """
+        + expectedNotification
+        + """
     Meldungsempfänger
     Name Bezirksamt Lichtenberg von Berlin | Gesundheitsamt
     Adresse Alfred-Kowalke-Straße 24, 10360 Berlin
@@ -187,11 +348,9 @@ class LaboratoryReportControllerIntegrationTest {
     Adresse 534 Erewhon St, 3999- PleasantVille
     Kontakt Telefon: 030 1234 (Dienstlich)
     Meldende Einrichtung
-    Name Primärlabor (Erregerdiagnostische Untersuchungsstelle)
-    Identifier BSNR: 98765430
-    DEMIS-Id: 13589
-    Adresse Dingsweg 321, 13055 Berlin, Deutschland
-    Kontakt Telefon: 0309876543210 (Dienstlich)
+    """
+        + expectedNotifierFacility
+        + """
     Betroffene Person
     Name Maxime Mustermann
     Geschlecht Weiblich
@@ -199,25 +358,34 @@ class LaboratoryReportControllerIntegrationTest {
     Adresse (Hauptwohnsitz) Teststrasse 123, 13055 Berlin, Deutschland
     Adresse (Derzeitiger Aufenthaltsort) Friedrichstraße 987, 66666 Berlin, Deutschland
     Kontakt Telefon: 030 555 (Privat)
+    """
+        + (expectedSubmitterFacility != null
+            ? expectedSubmitterFacility
+            : """
     Einsendende Person
     Name Dr. Dr. Otto Muster
     Adresse Strasse 123, 33445 Stuttgart
     Kontakt Fax: 030 1111 (Dienstlich)
+    """)
+        + """
     Laborbericht
     Erstellungs-/Bearbeitungszeitpunkt 04.03.2021 20:15
     Berichtsstatus Final
     Meldetatbestand SARS-CoV-2
     Ergebnis (Zusammenfassung) Meldepflichtiger Erreger nachgewiesen
     Erläuterung Ich bin die textuelle Conclusion ...
-    Auftragsnummer (E2E-Referenz) 2021-000672922
+    """
+        + getLaboratoryOrderNumber()
+        + "\n"
+        + """
     Test - SARS-CoV-2 (COVID-19) RNA [Presence] in Serum or Plasma by NAA with probe detection
     Teststatus Final
     Ergebniswert Detected
     Befund Positive
     Nachweismethode Nucleic acid assay (procedure)
     Erläuterung (Test) Nette Zusatzinformation …
-    Probeneingang 04.03.2021
-    Probenentnahme 04.03.2021
+    Probeneingang 14.05.2020
+    Probenentnahme 13.05.2020
     Probenmaterialstatus Verfügbar
     Probenmaterial Upper respiratory swab sample (specimen)
     Erläuterung (Probe) Ich bin eine interessante Zusatzinformation ...
@@ -242,9 +410,12 @@ class LaboratoryReportControllerIntegrationTest {
     Adresse Dingsweg 321, 13055 Berlin, Deutschland\s
     Kontakt Telefon: 0309876543210 (Dienstlich)
     e8d8cc43-32c2-4f93-8eaf-b2f3e6deb2a9
-    Auftragsnummer (E2E-Referenz) 2021-000672922
+    """
+        + getLaboratoryOrderNumber()
+        + "\n"
+        + """
     Meldetatbestand CVDP
-    Probenentnahme 04.03.2021\s
+    Probenentnahme 13.05.2020\s
     Probenmaterial Upper respiratory swab sample (specimen)
     Bitte ergänzen Sie den auf dem Probenbegleitschein genannten Labor-Identifikator für die Probe:""";
   }
@@ -288,5 +459,60 @@ class LaboratoryReportControllerIntegrationTest {
 
   private String cleanupString(final String input) {
     return input.replaceAll("\r\n", "\n").replaceAll("^\\s+", "").replaceAll("\\s+$", "");
+  }
+
+  private String getExpectedNotifierFacilityPdfText(final String contactPersonEntryString) {
+    return (featureFlags.isPdfOptimization() ? contactPersonEntryString + "\n" : "")
+        + """
+    Name Primärlabor (Erregerdiagnostische Untersuchungsstelle)
+    Identifier BSNR: 98765430
+    DEMIS-Id: 13589
+    Adresse Dingsweg 321, 13055 Berlin, Deutschland
+    Kontakt Telefon: 0309876543210 (Dienstlich)
+    """
+        + (!featureFlags.isPdfOptimization() ? contactPersonEntryString + "\n" : "");
+  }
+
+  private String getExpectedSubmitterFacilityPdfText(final String contactPersonEntryString) {
+    return """
+    Einsendende Einrichtung
+    """
+        + (featureFlags.isPdfOptimization() ? contactPersonEntryString + "\n" : "")
+        + """
+    Name Einsendepraxis ABC
+    Identifier BSNR: 135896780
+    Adresse Teststr. 123, 13589 Berlin, Deutschland
+    Kontakt Telefon: 030 1358967890 (Dienstlich)
+    """
+        + (!featureFlags.isPdfOptimization() ? contactPersonEntryString + "\n" : "");
+  }
+
+  private String getExpectedNotificationPdfText(String initalNotificationIdEntry) {
+    initalNotificationIdEntry =
+        !initalNotificationIdEntry.isEmpty() ? initalNotificationIdEntry + "\n" : "";
+    return (featureFlags.isPdfOptimization()
+        ? """
+    Meldungs-ID e8d8cc43-32c2-4f93-8eaf-b2f3e6deb2a9
+    """
+            + initalNotificationIdEntry
+            + """
+    Meldevorgangs-ID a5e00874-bb26-45ac-8eea-0bde76456703
+    Meldungserstellung 24.10.2023 09:06
+    Meldungsstatus Final
+    """
+        : """
+    Meldevorgangs-ID a5e00874-bb26-45ac-8eea-0bde76456703
+    Zeitpunkt des Eingangs 24.10.2023 09:06
+    Meldungs-ID e8d8cc43-32c2-4f93-8eaf-b2f3e6deb2a9
+    Meldungsstatus Final
+    Meldungserstellung/-änderung 04.03.2021 20:16
+    """
+            + initalNotificationIdEntry);
+  }
+
+  private String getLaboratoryOrderNumber() {
+    return (featureFlags.isPdfOptimization()
+        ? "Laboreigene Auftragsnummer 2021-000672922"
+        : "Auftragsnummer (E2E-Referenz) 2021-000672922");
   }
 }